move args parsing to common file to avoid import order mess and fix scatter tests
